+++ test-run-security/sec-groovy-gradle/server/src/main/groovy/com/sec/secgroovygradle/KotlinApplication.groovy
@@ -1,0 +1,13 @@
+package com.sec.secgroovygradle
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication
+class KotlinApplication {
+
+	static void main(String[] args) {
+		SpringApplication.run(KotlinApplication, args)
+	}
+
+}