+++ test-run-security/sec-groovy-gradle/server/src/main/groovy/com/sec/secgroovygradle/repository/UserRepository.groovy
@@ -1,0 +1,10 @@
+package com.sec.secgroovygradle.repository
+
+import com.sec.secgroovygradle.model.User
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface UserRepository extends JpaRepository<User, Long> {
+    User findByEmail(String email)
+}