+++ test-run-security/sec-groovy-gradle/server/src/test/groovy/com/sec/secgroovygradle/KotlinApplicationTests.groovy
@@ -1,0 +1,13 @@
+package com.sec.secgroovygradle
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class KotlinApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}