+++ test-run-security/sec-groovy-maven/server/src/main/groovy/com/sec/secgroovymaven/model/User.groovy
@@ -1,0 +1,21 @@
+package com.sec.secgroovymaven.model
+
+import jakarta.persistence.*
+import groovy.transform.ToString
+import groovy.transform.EqualsAndHashCode
+
+@Entity
+@Table(name = "users")
+@ToString
+@EqualsAndHashCode
+class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Long id
+    
+    @Column(nullable = false)
+    String name
+    
+    @Column(unique = true, nullable = false)
+    String email
+}