+++ test-run-security/sec-kotlin-gradle/server/src/main/kotlin/com/sec/seckotlingradle/KotlinApplication.kt
@@ -1,0 +1,11 @@
+package com.sec.seckotlingradle
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class KotlinApplication
+
+fun main(args: Array<String>) {
+	runApplication<KotlinApplication>(*args)
+}