+++ test-run-security/sec-kotlin-gradle/server/src/test/kotlin/com/sec/seckotlingradle/KotlinApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.sec.seckotlingradle
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class KotlinApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}